feature: calculate and display score and points for participants (#17)

+++ backend/Cbo.API/Models/DTO/GetMatchParticipantDto.cs
@@ -5,5 +5,5 @@ public record GetMatchParticipantDto
     public required int Id { get; set; }
     public required string Username { get; set; }
     public int? ScoreSum { get; set; }
-    public int? PointsSum { get; set; }
+    public decimal? PointsSum { get; set; }
 }
