+++ backend/Cbo.API/Models/DTO/GetTournamentParticipantDto.cs
@@ -8,7 +8,7 @@ public record GetTournamentParticipantDto
     public required string Username { get; set; }
     public required TournamentParticipantRole Role { get; set; }
     public int? ScoreSum { get; set; }
-    public int? PointsSum { get; set; }
+    public decimal? PointsSum { get; set; }
     public required int TournamentId { get; set; }
     public required int ApplicationUserId { get; set; }
 }
